+++ prototype/src/main/java/com/wantensoup/prototype/ScheduleDate/ScheduleDateRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScheduleDateRepository extends JpaRepository<ScheduleDate, Long> {
+public interface ScheduleDateRepository extends JpaRepository<ScheduleDate, Integer> {
 
     public ScheduleDate findByDate(String _date);
